Update Neptune_2025-11-13
All boxes shipping fromLocal used the old corePass logic. What's next: implementing the fromP1 and fromP2 gating logic (see if can re-use gateOK). For backlog and new sales, we assume that 40% of the total boxes are seats. We need to compare with the ending stock for that box (B2?) of P1 to see if we can ship all the boxes or not. Still need to think how we'll do it for P2